+++ app/src/main/java/de/live/gdev/timetracker/util/AppSettings.java
@@ -2,111 +2,140 @@ package de.live.gdev.timetracker.util;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.ColorRes;
-import android.support.v4.content.ContextCompat;
+import android.text.TextUtils;
 
+import de.live.gdev.timetracker.App;
 import de.live.gdev.timetracker.BuildConfig;
 import de.live.gdev.timetracker.R;
+import io.github.gsantner.opoc.util.AppSettingsBase;
 
-public class AppSettings {
-    private final SharedPreferences prefApp;
-    private final Context context;
+public class AppSettings extends AppSettingsBase {
+    private SharedPreferences prefCurrentProfile;
 
-    public AppSettings(Context context) {
-        this.context = context.getApplicationContext();
-        prefApp = this.context.getSharedPreferences("app", Context.MODE_PRIVATE);
+    //#####################
+    //## Methods
+    //#####################
+    private AppSettings(Context context) {
+        super(context);
+        loadSelectedProfile();
     }
 
-    public Context getApplicationContext() {
-        return context;
+    public static AppSettings get() {
+        return new AppSettings(App.get());
     }
 
-    public void clearAppSettings() {
-        prefApp.edit().clear().commit();
-    }
+    //###########################################
+    //## Settings options
+    //############################################
 
-    public String getKey(int stringKeyResourceId) {
-        return context.getString(stringKeyResourceId);
-    }
-
-    public boolean isKeyEqual(String key, int stringKeyRessourceId) {
-        return key.equals(getKey(stringKeyRessourceId));
-    }
-
-    private void setString(SharedPreferences pref, int keyRessourceId, String value) {
-        pref.edit().putString(context.getString(keyRessourceId), value).apply();
-    }
-
-    private void setInt(SharedPreferences pref, int keyRessourceId, int value) {
-        pref.edit().putInt(context.getString(keyRessourceId), value).apply();
-    }
-
-    private void setLong(SharedPreferences pref, int keyRessourceId, long value) {
-        pref.edit().putLong(context.getString(keyRessourceId), value).apply();
-    }
-
-    private void setBool(SharedPreferences pref, int keyRessourceId, boolean value) {
-        pref.edit().putBoolean(context.getString(keyRessourceId), value).apply();
-    }
-
-    private void setStringArray(SharedPreferences pref, int keyRessourceId, Object[] values) {
-        StringBuilder sb = new StringBuilder();
-        for (Object value : values) {
-            sb.append("%%%");
-            sb.append(value.toString());
-        }
-        setString(pref, keyRessourceId, sb.toString().replaceFirst("%%%", ""));
-    }
-
-    private String[] getStringArray(SharedPreferences pref, int keyRessourceId) {
-        String value = pref.getString(context.getString(keyRessourceId), "%%%");
-        if (value.equals("%%%")) {
-            return new String[0];
-        }
-        return value.split("%%%");
-    }
-
-    private String getString(SharedPreferences pref, int ressourceId, String defaultValue) {
-        return pref.getString(context.getString(ressourceId), defaultValue);
-    }
-
-    private String getString(SharedPreferences pref, int ressourceId, int ressourceIdDefaultValue) {
-        return pref.getString(context.getString(ressourceId), context.getString(ressourceIdDefaultValue));
-    }
-
-    private boolean getBool(SharedPreferences pref, int ressourceId, boolean defaultValue) {
-        return pref.getBoolean(context.getString(ressourceId), defaultValue);
-    }
-
-    private int getInt(SharedPreferences pref, int ressourceId, int defaultValue) {
-        return pref.getInt(context.getString(ressourceId), defaultValue);
-    }
-
-    private long getLong(SharedPreferences pref, int ressourceId, long defaultValue) {
-        return pref.getLong(context.getString(ressourceId), defaultValue);
-    }
-
-
-    public int getColor(SharedPreferences pref, String key, int defaultColor) {
-        return pref.getInt(key, defaultColor);
-    }
-
-    public int getColorRes(@ColorRes int resColorId) {
-        return ContextCompat.getColor(context, resColorId);
-    }
-
-    /*
-    //     Setters & Getters
-    */
     public boolean isAppFirstStart() {
         boolean value = getBool(prefApp, R.string.pref_key__app_first_start, true);
         setBool(prefApp, R.string.pref_key__app_first_start, false);
         return value;
     }
 
+    @SuppressWarnings("ConstantConditions")
     public boolean isAppCurrentVersionFirstStart() {
         int value = getInt(prefApp, R.string.pref_key__app_first_start_current_version, -1);
         setInt(prefApp, R.string.pref_key__app_first_start_current_version, BuildConfig.VERSION_CODE);
         return value != BuildConfig.VERSION_CODE && !BuildConfig.IS_TEST_BUILD;
+    }
+
+    public boolean isShowMainFab() {
+        return getBool(prefApp, R.string.pref_key__show_main_fab, true);
+    }
+
+    public boolean isReloadRequired() {
+        return getBool(prefApp, R.string.pref_key__app_reload_required, false);
+    }
+
+    public void setReloadRequired(boolean value) {
+        setBool(prefApp, R.string.pref_key__app_reload_required, value);
+    }
+
+    public SharedPreferences getSharedPreferenceCurrentProfile() {
+        return prefCurrentProfile;
+    }
+
+
+    //#############
+    //# Profile
+    //#############
+    public void loadSelectedProfile() {
+        prefCurrentProfile = context.getSharedPreferences("Profile" + getSelectedProfileNr(), Context.MODE_PRIVATE);
+    }
+
+    public void loadProfile(int nr) {
+        nr = nr >= 0 && nr < 3 ? nr : 0;
+        prefCurrentProfile = context.getSharedPreferences("Profile" + nr, Context.MODE_PRIVATE);
+    }
+
+    public int getSelectedProfileNr() {
+        return getIntOfStringPref(R.string.pref_key__app_selected_profile, 0);
+    }
+
+    public boolean isProfileEmpty() {
+        return getProfilePathDomainAndDirectory().equals(rstr(R.string.default_path_domain_and_folder))
+                || TextUtils.isEmpty(getProfileLoginPassword());
+    }
+
+    public String getProfileSummary() {
+        return isProfileEmpty() ? rstr(R.string.no_data)
+                : String.format("%s @%s", getProfileLoginUsername(), getProfilePathDomainAndDirectory());
+    }
+
+    public String getProfilePathDomainAndDirectory() {
+        String ret = getString(prefCurrentProfile, R.string.pref_key__profile_path_domain_and_directory, rstr(R.string.default_path_domain_and_folder));
+        if (!TextUtils.isEmpty(ret) && ret.endsWith("/")) {
+            ret = ret.substring(0, ret.length() - 1);
+            setProfilePathDomainAndDirectory(ret);
+        }
+        return ret;
+    }
+
+    public String getProfilePathFull() {
+        String ret = getProfilePathDomainAndDirectory();
+        ret = TextUtils.isEmpty(getProfilePathFilename()) ? ret : ret + "/" + getProfilePathFilename();
+        return ret;
+    }
+
+    public void setProfilePathDomainAndDirectory(String value) {
+        setString(prefCurrentProfile, R.string.pref_key__profile_path_domain_and_directory, value);
+    }
+
+    public String getProfilePathFilename() {
+        return getString(prefCurrentProfile, R.string.pref_key__profile_path_filename, "index.php");
+    }
+
+    public String getProfileLoginUsername() {
+        return getString(prefCurrentProfile, R.string.pref_key__profile_login_username, rstr(R.string.default_username));
+    }
+
+    public String getProfileLoginPassword() {
+        return getString(prefCurrentProfile, R.string.pref_key__profile_login_password, "");
+    }
+
+    public boolean isProfileAutoLogin() {
+        return getBool(prefCurrentProfile, R.string.pref_key__profile_autologin, true);
+    }
+
+    public boolean isProfileAcceptAllSsl() {
+        return getBool(prefCurrentProfile, R.string.pref_key__profile_accept_all_ssl_certs, false);
+    }
+
+    public boolean isProfileLoadInDesktopMode() {
+        return getBool(prefCurrentProfile, R.string.pref_key__profile_load_in_desktop_mode, true);
+    }
+
+    public void setProfileLoadInDesktopMode(boolean value) {
+        setBool(prefCurrentProfile, R.string.pref_key__profile_load_in_desktop_mode, value);
+    }
+
+    public String getProfileHttpAuthUsername() {
+        return getString(prefCurrentProfile, R.string.pref_key__profile_httpauth_username, rstr(R.string.default_username));
+    }
+
+    public String getProfileHttpAuthPassword() {
+        return getString(prefCurrentProfile, R.string.pref_key__profile_httpauth_password, "");
     }
 }